24-bit input

+++ include/faac.h
@@ -16,7 +16,7 @@
  * License along with this library; if not, write to the Free Software
  * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
  *
- * $Id: faac.h,v 1.27 2003/06/26 19:18:02 knik Exp $
+ * $Id: faac.h,v 1.28 2003/07/10 19:16:01 knik Exp $
  */
 
 #ifndef FAACLIB_H
@@ -117,11 +117,11 @@ int FAACAPI faacEncGetDecoderSpecificInfo(faacEncHandle hEncoder,
 										  unsigned long* pSizeOfDecoderSpecificInfo);
 
 int FAACAPI faacEncEncode(faacEncHandle hEncoder,
-						  short *inputBuffer,
-						  unsigned int samplesInput,
-						  unsigned char *outputBuffer,
-						  unsigned int bufferSize
-						  );
+			  int32_t *inputBuffer,
+			  unsigned int samplesInput,
+			  unsigned char *outputBuffer,
+			  unsigned int bufferSize
+			 );
 
 int FAACAPI faacEncClose(faacEncHandle hEncoder);
